feat(database): ✨ added sqlite extension for compression + cleaned few things
Moved sheikh db stuff to iqra for now, changed name of scraper to sheikh
