+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["./crates/iqra", "./crates/sheikh"]
+members = ["./crates/iqra"]
 resolver = "2"
