+++ apps/salah/package.json
@@ -14,11 +14,11 @@
 	},
 	"devDependencies": {
 		"@quran/tsconfig": "workspace:^0.0.0",
+		"@typescript-eslint/eslint-plugin": "^5.47.0",
+		"@typescript-eslint/parser": "^5.47.0",
 		"autoprefixer": "^10.4.13",
 		"eslint": "^8.30.0",
 		"eslint-config-custom": "workspace:*",
-		"@typescript-eslint/eslint-plugin": "^5.47.0",
-		"@typescript-eslint/parser": "^5.47.0",
 		"eslint-plugin-import": "^2.26.0",
 		"eslint-plugin-jsx-a11y": "^6.6.1",
 		"eslint-plugin-simple-import-sort": "^8.0.0",
