+++ apps/tanzil/package.json
@@ -22,8 +22,8 @@
 		"eslint": "^8.30.0",
 		"eslint-config-custom": "workspace:*",
 		"eslint-plugin-import": "^2.26.0",
-		"eslint-plugin-simple-import-sort": "^8.0.0",
 		"eslint-plugin-jsx-a11y": "^6.6.1",
+		"eslint-plugin-simple-import-sort": "^8.0.0",
 		"postcss": "^8.4.20",
 		"tailwindcss": "^3.2.4",
 		"typescript": "^4.9.4",
