+++ crates/iqra/Cargo.toml
@@ -15,9 +15,18 @@ description = "A Tauri App"
 tauri-build = { version = "1.2.1", features = [] }
 
 [dependencies]
+anyhow = '1'
+home = '0.5'
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+sqlx = { version = '0.6', features = ['runtime-tokio-rustls', 'sqlite'] }
 tauri = { version = "1.2.2", features = [] }
+tokio = { version = '1', features = ['full'] }
+tracing = '0.1'
+tracing-appender = '0.2'
+tracing-bunyan-formatter = '0.3'
+tracing-log = '0.1'
+tracing-subscriber = { version = '0.3', features = ['env-filter', 'std'] }
 
 [features]
 # by default Tauri runs in production mode
