+++ crates/iqra/src/main.rs
@@ -2,9 +2,30 @@
     all(not(debug_assertions), target_os = "windows"),
     windows_subsystem = "windows"
 )]
+use anyhow::{Ok, Result};
+use tracing_bunyan_formatter::{BunyanFormattingLayer, JsonStorageLayer};
+use tracing_log::LogTracer;
+use tracing_subscriber::{filter::EnvFilter, layer::SubscriberExt, Registry};
 
-fn main() {
+use crate::db::setup_db;
+
+mod db;
+
+#[tokio::main]
+async fn main() -> Result<()> {
+    LogTracer::init().expect("Unable to setup log tracer!");
+    let app_name = concat!(env!("CARGO_PKG_NAME"), "-", env!("CARGO_PKG_VERSION")).to_string();
+    let (non_blocking_writer, _guard) = tracing_appender::non_blocking(std::io::stdout());
+    let bunyan_formatting_layer = BunyanFormattingLayer::new(app_name, non_blocking_writer);
+    let subscriber = Registry::default()
+        .with(EnvFilter::new("INFO"))
+        .with(JsonStorageLayer)
+        .with(bunyan_formatting_layer);
+
+    tracing::subscriber::set_global_default(subscriber).unwrap();
+    setup_db().await?;
     tauri::Builder::default()
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
+    Ok(())
 }
