--- crates/sheikh/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = 'sheikh'
-version = '0.1.0'
-edition = '2021'
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = '1'
-home = '0.5'
-serde = { version = '1.0', features = ['derive'] }
-sqlx = { version = '0.6', features = ['runtime-tokio-rustls', 'sqlite'] }
-tokio = { version = '1', features = ['full'] }
-tracing = '0.1'
-tracing-appender = '0.2'
-tracing-bunyan-formatter = '0.3'
-tracing-log = '0.1'
-tracing-subscriber = { version = '0.3', features = ['env-filter', 'std'] }
