--- crates/sheikh/src/db.rs
@@ -1,9 +1,0 @@
-use anyhow::{Ok, Result};
-use sqlx::{sqlite::SqlitePool, Pool, Sqlite};
-use tracing::info;
-
-pub async fn setup_db() -> Result<Pool<Sqlite>> {
-    let pool = SqlitePool::connect("sqlite:./crates/sheikh/src/dev.db").await?;
-    info!("DB SETUP SUCCESSFUL");
-    Ok(pool)
-}
