+++ package.json
@@ -10,13 +10,13 @@
 		"dev": "turbo run dev",
 		"format": "dprint fmt",
 		"lint": "turbo run lint",
-		"prepare": "husky install",
-		"tanzil": "pnpm --filter=./apps/tanzil",
 		"mushaf": "pnpm --filter=./apps/mushaf",
+		"prepare": "husky install",
 		"salah": "pnpm --filter=./apps/salah",
-		"tarteel": "pnpm --filter=./apps/tarteel",
+		"sheikh": "pnpm --filter=./packages/sheikh",
+		"tanzil": "pnpm --filter=./apps/tanzil",
 		"tanzil:dev": "cargo tauri dev",
-		"scraper": "pnpm --filter=./packages/scraper",
+		"tarteel": "pnpm --filter=./apps/tarteel",
 		"test": "turbo run test"
 	},
 	"devDependencies": {
@@ -24,5 +24,5 @@
 		"husky": "^8.0.2",
 		"turbo": "^1.6.3"
 	},
-	"packageManager": "pnpm@7.19.0"
+	"packageManager": "pnpm@7.20.0"
 }
