--- packages/scraper/tsconfig.json
@@ -1,8 +1,0 @@
-{
-	"extends": "@quran/tsconfig/simple.json",
-	"include": ["**/*.ts", "**/*.tsx"],
-	"exclude": ["node_modules"],
-	"ts-node": {
-		"swc": true
-	}
-}
