+++ packages/sheikh/.eslintrc.cjs
@@ -13,4 +13,8 @@ module.exports = {
 			},
 		},
 	},
+	rules: {
+		'no-console': 'off',
+		'import/no-named-as-default': 'off',
+	},
 };
