+++ packages/sheikh/package.json
@@ -1,9 +1,20 @@
 {
-	"name": "@quran/scraper",
+	"name": "@quran/sheikh",
+	"version": "1.0.0",
+	"main": "src/index.ts",
+	"scripts": {
+		"dev": "ts-node-esm src/index.ts",
+		"lint": "eslint . --ext .ts,.tsx"
+	},
+	"dependencies": {
+		"better-sqlite3": "^8.0.1",
+		"kysely": "^0.23.0"
+	},
 	"devDependencies": {
 		"@quran/tsconfig": "workspace:^0.0.0",
 		"@swc/core": "^1.3.24",
 		"@swc/helpers": "^0.4.14",
+		"@types/better-sqlite3": "^7.6.3",
 		"@types/node": "^18.11.17",
 		"@typescript-eslint/eslint-plugin": "^5.47.0",
 		"@typescript-eslint/parser": "^5.47.0",
@@ -16,10 +27,5 @@
 		"tslib": "^2.4.1",
 		"typescript": "^4.9.4"
 	},
-	"main": "src/index.ts",
-	"scripts": {
-		"dev": "ts-node src/index.ts",
-		"lint": "eslint . --ext .ts,.tsx"
-	},
-	"version": "1.0.0"
+	"type": "module"
 }
