+++ packages/sheikh/tsconfig.json
@@ -1,0 +1,15 @@
+{
+	"extends": "@quran/tsconfig/simple.json",
+	"include": ["**/*.ts", "**/*.tsx"],
+	"exclude": ["node_modules"],
+	"ts-node": {
+		"swc": true,
+		"esm": true
+	},
+	"compilerOptions": {
+		"module": "NodeNext",
+		"moduleResolution": "NodeNext",
+		"allowSyntheticDefaultImports": true,
+		"lib": ["ES2022"]
+	}
+}
